# contracts/openapi/domain-services.v1.yaml
openapi: 3.1.0
info:
  title: SurakshaSetu domain services
  version: 1.0.0
  description: |
    The tier contract between the orchestrator (Python) and the domain services (Java). It is the
    single source of truth: Java interfaces and models and Python models are generated from it.
    Model shapes follow the TDD (§3.5–§3.8) where it defines them.

    ## Conventions
    - **Money** is a decimal string of rupees (`^\d+(\.\d{1,2})?$`), never a JSON number.
    - **Dates** are ISO 8601 (`date` or `date-time`); date-times are UTC.
    - **Errors** are `application/problem+json` with `type`, `title`, `status`, `detail` and a
      machine-readable `code`.
    - **Auth**: a bearer token (a shared dev token locally; mTLS in production).
    - **Decisions** (`POST` eligibility, suitability, ranking and quotes): the request carries
      `pins.rules` and the service evaluates exactly that version, or answers 409
      `RULES_VERSION_UNKNOWN`. The response carries `decision_id`, the version it used
      (`rules_version`, `ranker_version` or `rating_version`), `inputs_sha256` and `reason_codes`.
    - Commission, margin, incentive, campaign and payout fields never appear in this contract.

    ## Problem codes
    | Status | code | When |
    | --- | --- | --- |
    | 404 | `NOT_FOUND` | Unknown UIN, consent id, notice, disclosure, pincode or occupation. |
    | 409 | `RULES_VERSION_UNKNOWN` | `pins.rules` names a version the service has not loaded. |
...

    ## Hashing
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/disclosures/sets/{uin}:
    get:
      tags: [disclosure]
      operationId: getDisclosureSet
      summary: The ordered disclosure set for a product, channel and language.
      parameters:
        - $ref: '#/components/parameters/Uin'
        - $ref: '#/components/parameters/Channel'
        - $ref: '#/components/parameters/Language'
        - $ref: '#/components/parameters/AsOf'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DisclosureSet'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/disclosures/{disclosure_id}:
    get:
      tags: [disclosure]
      operationId: getDisclosure
      summary: One disclosure, e.g. the global items DISC-GLOBAL-AI-06, -QUOTE-02 and -TAX-05.
      parameters:
        - name: disclosure_id
          in: path
          required: true
          schema: {type: string}
        - $ref: '#/components/parameters/Language'
        - $ref: '#/components/parameters/AsOf'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Disclosure'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/reference/pincodes/{pincode}:
    get:
      tags: [reference]
      operationId: getPincode
      parameters:
        - name: pincode
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Pincode'}
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PincodeInfo'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/reference/occupations:
    get:
      tags: [reference]
      operationId: searchOccupations
      parameters:
        - name: q
          in: query
          description: Case-insensitive search over code and label.
          schema: {type: string}
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Occupation'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/reference/occupations/{code}:
    get:
      tags: [reference]
      operationId: getOccupation
      parameters:
        - name: code
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Occupation'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Language:
      name: language
      in: query
      required: true
      schema: {type: string}
      example: en-IN
    Channel:
      name: channel
      in: query
      required: true
      schema: {type: string}
      example: web
    AsOf:
      name: as_of
      in: query
      description: Evaluate as of this instant. Defaults to now.
      schema: {type: string, format: date-time}
    Uin:
      name: uin
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uin'}
...
  responses:
    Problem:
      description: Error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Problem:
      type: object
      required: [title, status, code]
      properties:
        type: {type: string, default: 'about:blank', description: 'RFC 9457: absent means about:blank.'}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance:
          type: string
          description: 'RFC 9457: the request path. It can carry ids or a pincode, so never log it.'
        code:
          type: string
          description: Machine-readable error code; the table in the API description lists them.
...
    Sha256Hex:
      type: string
      pattern: '^[0-9a-f]{64}$'
    Uin:
      type: string
      description: IRDAI unique identification number, e.g. 999N001V02 (product) or 999A007V01 (rider).
      pattern: '^[0-9]{3}[A-Z][0-9]{3}V[0-9]{2}$'
    Pincode:
      type: string
      pattern: '^[1-9][0-9]{5}$'
...
    DisclosureItem:
      type: object
      required: [disclosure_id, body, body_sha256]
      properties:
        disclosure_id: {type: string, examples: [DISC-GLOBAL-SOLICIT-01]}
        body: {type: string, description: Verbatim approved text.}
        body_sha256: {$ref: '#/components/schemas/Sha256Hex'}
    DisclosureSet:
      type: object
      required: [uin, channel, language, registry_version, items, set_sha256, is_dummy]
      properties:
        uin: {$ref: '#/components/schemas/Uin'}
        channel: {type: string}
        language: {type: string}
        registry_version: {type: string}
        items:
          type: array
          description: In the order they must be shown.
          items: {$ref: '#/components/schemas/DisclosureItem'}
        set_sha256: {$ref: '#/components/schemas/Sha256Hex'}
        is_dummy: {type: boolean}
    Disclosure:
      type: object
      required: [disclosure_id, language, body, body_sha256, is_dummy]
      properties:
        disclosure_id: {type: string}
        language: {type: string}
        body: {type: string, description: Verbatim approved text.}
        body_sha256: {$ref: '#/components/schemas/Sha256Hex'}
        is_dummy: {type: boolean}

    # --- Reference --------------------------------------------------------------------------
    PincodeInfo:
      type: object
      required: [pincode, district, state, serviceable, is_dummy]
      properties:
        pincode: {$ref: '#/components/schemas/Pincode'}
        district: {type: string}
        state: {type: string}
        serviceable: {type: boolean}
        is_dummy: {type: boolean}
    Occupation:
      type: object
      required: [code, label, risk_class, is_dummy]
      properties:
        code: {type: string}
        label: {type: string}
        risk_class: {type: integer, minimum: 1, maximum: 4}
        is_dummy: {type: boolean}

    # --- Eligibility (TDD §3.6) -------------------------------------------------------------
...